Added the pre-processing section of the code.
I merged all of our pre-processing parts together. For the normalization part, I used the mean and standard deviation from the training set only, to normalize both sets. I also included the code to apply the standard deviation and mean from the testing set to normalize the testing set, but it is commented out for now. Once we decide which is the correct version, I will put the code in a .py file.