This code generates the data without capped prices
The capped values of prices USD 500000 might affect the performance of modeling. Referring to Part1 EDA, prices have a set of capped values (median_price_value histogram). So to check the effect of these on models and see if removing these values will improve the models' performance the values of prices >USD 490000 are removed in this code. Test this dataset on your models and compare the results.