Categorical feature (ocean proximity) update
In this code, the categorical feature (ocean proximity) is updated to assign numerical values to ocean proximity in the gradient order: the lower the number the further away is the house from the ocean. This gradient is chosen for the better and easier interpretation of models results (feature importances). USE THIS CODE for models now.